requirements.txt
django
djangorestframework
django-filter
django-environ
django-cors-headers
django-autoslug
bcrypt
